+++ src/internal/processProp.ts
@@ -5,16 +5,17 @@ import type {
   DecoratedPropertyMetadata,
   DiscriminatorObject,
   KeyStringAny,
+  MappedInnerOuterOptions,
   NestedDiscriminatorsMap,
   VirtualPopulateMap,
 } from '../types';
-import { DecoratorKeys, WhatIsIt } from './constants';
+import { DecoratorKeys, PropType } from './constants';
 import { schemas } from './data';
 import {
   CannotBeSymbolError,
   InvalidEnumTypeError,
   InvalidTypeError,
-  InvalidWhatIsItError,
+  InvalidPropTypeError,
   NotAllVPOPElementsError,
   NotNumberTypeError,
   NotStringTypeError,
@@ -34,7 +35,7 @@ export function processProp(input: DecoratedPropertyMetadata): void {
   const name = utils.getName(target);
   const rawOptions: KeyStringAny = Object.assign({}, input.options);
   let Type: any | undefined = Reflect.getMetadata(DecoratorKeys.Type, target, key);
-  const propKind = input.whatis ?? detectWhatIsIt(Type);
+  const propKind = input.whatis ?? detectPropType(Type);
 
   logger.debug('Starting to process "%s.%s"', name, key);
   utils.assertion(typeof key === 'string', () => new CannotBeSymbolError(name, key));
@@ -44,10 +45,10 @@ export function processProp(input: DecoratedPropertyMetadata): void {
   {
     // soft errors & "type"-alias mapping
     switch (propKind) {
-      case WhatIsIt.NONE:
+      case PropType.NONE:
         break;
-      case WhatIsIt.MAP:
-      case WhatIsIt.ARRAY:
+      case PropType.MAP:
+      case PropType.ARRAY:
         // set the "Type" to undefined if "ref" or "refPath" are defined, as an fallback in case "type" is also not defined
         if (('ref' in rawOptions || 'refPath' in rawOptions) && !('type' in rawOptions)) {
           Type = undefined;
@@ -79,15 +80,15 @@ export function processProp(input: DecoratedPropertyMetadata): void {
     Type = mongoose.Schema.Types.Buffer;
   }
 
-  // confirm that "WhatIsIt" is an ARRAY and if that the Type is still an *ARRAY, set them to Mixed
+  // confirm that "PropType" is an ARRAY and if that the Type is still an *ARRAY, set them to Mixed
   // for issues like https://github.com/typegoose/typegoose/issues/300
-  if (propKind === WhatIsIt.ARRAY && detectWhatIsIt(Type) === WhatIsIt.ARRAY) {
+  if (propKind === PropType.ARRAY && detectPropType(Type) === PropType.ARRAY) {
     logger.debug('Type is still *ARRAY, defaulting to Mixed');
     Type = mongoose.Schema.Types.Mixed;
   }
 
-  // confirm that "WhatIsIt" is an MAP and if that the Type is still an *MAP, set them to Mixed
-  if (propKind === WhatIsIt.MAP && detectWhatIsIt(Type) === WhatIsIt.MAP) {
+  // confirm that "PropType" is an MAP and if that the Type is still an *MAP, set them to Mixed
+  if (propKind === PropType.MAP && detectPropType(Type) === PropType.MAP) {
     logger.debug('Type is still *Map, defaulting to Mixed');
     Type = mongoose.Schema.Types.Mixed;
   }
@@ -171,11 +172,11 @@ export function processProp(input: DecoratedPropertyMetadata): void {
     }
 
     switch (propKind) {
-      case WhatIsIt.ARRAY:
+      case PropType.ARRAY:
         schemaProp[key] = utils.mapArrayOptions(rawOptions, newType, target, key);
 
         return;
-      case WhatIsIt.MAP:
+      case PropType.MAP:
         const mapped = utils.mapOptions(rawOptions, newType, target, key);
 
         schemaProp[key] = {
@@ -185,7 +186,7 @@ export function processProp(input: DecoratedPropertyMetadata): void {
         };
 
         return;
-      case WhatIsIt.NONE:
+      case PropType.NONE:
         schemaProp[key] = {
           ...rawOptions,
           type: newType,
@@ -193,7 +194,7 @@ export function processProp(input: DecoratedPropertyMetadata): void {
 
         return;
       default:
-        throw new InvalidWhatIsItError(propKind, name, key, 'whatis(Passthrough)');
+        throw new InvalidPropTypeError(propKind, name, key, 'PropType(Passthrough)');
     }
   }
 
@@ -205,17 +206,17 @@ export function processProp(input: DecoratedPropertyMetadata): void {
     delete rawOptions.ref;
 
     switch (propKind) {
-      case WhatIsIt.ARRAY:
+      case PropType.ARRAY:
         schemaProp[key] = utils.mapArrayOptions(rawOptions, refType, target, key, undefined, { ref });
         break;
-      case WhatIsIt.NONE:
+      case PropType.NONE:
         schemaProp[key] = {
           type: refType,
           ref,
           ...rawOptions,
         };
         break;
-      case WhatIsIt.MAP:
+      case PropType.MAP:
         const mapped = utils.mapOptions(rawOptions, refType, target, key);
 
         schemaProp[key] = {
@@ -229,7 +230,7 @@ export function processProp(input: DecoratedPropertyMetadata): void {
         };
         break;
       default:
-        throw new InvalidWhatIsItError(propKind, name, key, 'whatis(ref)');
+        throw new InvalidPropTypeError(propKind, name, key, 'PropType(ref)');
     }
 
     return;
@@ -245,10 +246,10 @@ export function processProp(input: DecoratedPropertyMetadata): void {
     );
 
     switch (propKind) {
-      case WhatIsIt.ARRAY:
+      case PropType.ARRAY:
         schemaProp[key] = utils.mapArrayOptions(rawOptions, refType, target, key, undefined, { refPath });
         break;
-      case WhatIsIt.NONE:
+      case PropType.NONE:
         schemaProp[key] = {
           type: refType,
           refPath,
@@ -256,7 +257,7 @@ export function processProp(input: DecoratedPropertyMetadata): void {
         };
         break;
       default:
-        throw new InvalidWhatIsItError(propKind, name, key, 'whatis(refPath)');
+        throw new InvalidPropTypeError(propKind, name, key, 'PropType(refPath)');
     }
 
     return;
@@ -360,21 +361,32 @@ export function processProp(input: DecoratedPropertyMetadata): void {
     }
 
     switch (propKind) {
-      case WhatIsIt.ARRAY:
+      case PropType.ARRAY:
         schemaProp[key] = utils.mapArrayOptions(rawOptions, Type, target, key);
 
         return;
-      case WhatIsIt.MAP:
-        const mapped = utils.mapOptions(rawOptions, Type, target, key);
+      case PropType.MAP:
+        let mapped: MappedInnerOuterOptions;
+        let finalType: mongoose.SchemaTypeOptions<any>;
+
+        // Map the correct options for the end type
+        if (utils.isTypeMeantToBeArray(rawOptions)) {
+          mapped = utils.mapOptions(rawOptions, mongoose.Schema.Types.Array, target, key);
+          // "rawOptions" is not used here, because that would duplicate some options to where the should not be
+          finalType = utils.mapArrayOptions({ ...mapped.inner, dim: rawOptions.dim }, Type, target, key);
+        } else {
+          mapped = utils.mapOptions(rawOptions, Type, target, key);
+          finalType = { ...mapped.inner, type: Type };
+        }
 
         schemaProp[key] = {
           ...mapped.outer,
           type: Map,
-          of: { type: Type, ...mapped.inner },
+          of: { ...finalType },
         };
 
         return;
-      case WhatIsIt.NONE:
+      case PropType.NONE:
         schemaProp[key] = {
           ...rawOptions,
           type: Type,
@@ -382,7 +394,7 @@ export function processProp(input: DecoratedPropertyMetadata): void {
 
         return;
       default:
-        throw new InvalidWhatIsItError(propKind, name, key, 'whatis(primitive)');
+        throw new InvalidPropTypeError(propKind, name, key, 'PropType(primitive)');
     }
   }
 
@@ -403,11 +415,11 @@ export function processProp(input: DecoratedPropertyMetadata): void {
 
   const virtualSchema = buildSchema(Type);
   switch (propKind) {
-    case WhatIsIt.ARRAY:
+    case PropType.ARRAY:
       schemaProp[key] = utils.mapArrayOptions(rawOptions, virtualSchema, target, key, Type);
 
       return;
-    case WhatIsIt.MAP:
+    case PropType.MAP:
       // special handling if the lower type should be an array
       if ('dim' in rawOptions) {
         logger.debug('Map SubDocument Array for "%s.%s"', name, key);
@@ -432,7 +444,7 @@ export function processProp(input: DecoratedPropertyMetadata): void {
       };
 
       return;
-    case WhatIsIt.NONE:
+    case PropType.NONE:
       schemaProp[key] = {
         ...rawOptions,
         type: virtualSchema,
@@ -440,7 +452,7 @@ export function processProp(input: DecoratedPropertyMetadata): void {
 
       return;
     default:
-      throw new InvalidWhatIsItError(propKind, name, key, 'whatis(subSchema)');
+      throw new InvalidPropTypeError(propKind, name, key, 'PropType(subSchema)');
   }
 }
 
@@ -452,11 +464,11 @@ export function processProp(input: DecoratedPropertyMetadata): void {
 // function optionDeprecation(options: any) {}
 
 /**
- * Detect "WhatIsIt" based on "Type"
+ * Detect "PropType" based on "Type"
  * @param Type The Type used for detection
  */
-function detectWhatIsIt(Type: any): WhatIsIt {
-  logger.debug('Detecting WhatIsIt');
+function detectPropType(Type: any): PropType {
+  logger.debug('Detecting PropType');
 
   if (
     Type === Array ||
@@ -465,11 +477,11 @@ function detectWhatIsIt(Type: any): WhatIsIt {
     Type === mongoose.Types.DocumentArray ||
     Type === mongoose.Schema.Types.DocumentArray
   ) {
-    return WhatIsIt.ARRAY;
+    return PropType.ARRAY;
   }
   if (Type === Map || Type === mongoose.Types.Map || Type === mongoose.Schema.Types.Map) {
-    return WhatIsIt.MAP;
+    return PropType.MAP;
   }
 
-  return WhatIsIt.NONE;
+  return PropType.NONE;
 }