+++ src/queryMethod.ts
@@ -2,7 +2,7 @@ import type { Query } from 'mongoose';
 import { DecoratorKeys } from './internal/constants';
 import { getName } from './internal/utils';
 import { logger } from './logSettings';
-import type { AnyParamConstructor, QueryMethodMap, ReturnModelType } from './types';
+import type { AnyParamConstructor, QueryHelperThis, QueryMethodMap } from './types';
 
 /**
  * Adds a query method to schema.
@@ -28,7 +28,7 @@ import type { AnyParamConstructor, QueryMethodMap, ReturnModelType } from './typ
  * ```
  */
 export function queryMethod<QueryHelpers, U extends AnyParamConstructor<any>>(
-  func: (this: ReturnModelType<U, QueryHelpers>, ...params: any[]) => Query<any, any>
+  func: (this: QueryHelperThis<U, QueryHelpers>, ...params: any[]) => Query<any, any>
 ): ClassDecorator {
   return (target: any) => {
     logger.info('Adding query method "%s" to %s', func.name, getName(target));