+++ test/models/typeguards.ts
@@ -1,5 +1,4 @@
-import { WhatIsIt } from '../../src/internal/constants';
-import { getModelForClass, mongoose, prop, Ref } from '../../src/typegoose';
+import { getModelForClass, mongoose, prop, Ref, PropType } from '../../src/typegoose';
 
 export class IsRefTypeNestedString {
   @prop()
@@ -39,7 +38,7 @@ export class Sub {
 }
 
 export class MTypesArrayRef {
-  @prop({ required: true, ref: () => Sub }, WhatIsIt.ARRAY)
+  @prop({ required: true, ref: () => Sub }, PropType.ARRAY)
   public subs!: mongoose.Types.Array<Ref<Sub>>;
 }